fix humor change on game

+++ Assets/Modules/ActorModule/Scripts/Presentation/HumorBarView.cs
@@ -6,7 +6,7 @@ namespace Modules.ActorModule.Scripts.Presentation
     {
         event Action OnViewEnable;
         event Action OnViewDisable;
-        void HumorChange(int humorLevel, int lastHumorChange, int maxHumor);
-        void InitView(int humor, int maxHumor);
+        void HumorChange(float humorLevel, float lastHumorChange, float maxHumor);
+        void InitView(float humor, float maxHumor);
     }
 }
